+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 12.35.0-beta.2 (Unreleased)
+## 12.36.0-beta.1 (Unreleased)
 
 ### Features Added
 
@@ -17,6 +17,21 @@
   logs a warning and signals end-of-file when such an error occurs at or past the known end of the resource.
 
 ### Other Changes
+
+## 12.35.0 (2026-06-11)
+
+### Features Added
+- Added support for Blob Smart Tier.
+- Added support for virtual directory-scoped SAS.
+- Added support for service version 2026-06-06.
+
+### Other Changes
+
+#### Dependency Updates
+- Upgraded `azure-storage-common` from `12.33.0` to version `12.34.0`.
+- Upgraded `azure-storage-internal-avro` from `12.19.0` to version `12.20.0`.
+- Upgraded `azure-core` from `1.58.0` to version `1.58.1`.
+- Upgraded `azure-core-http-netty` from `1.16.4` to version `1.16.5`.
 
 ## 12.34.0 (2026-05-14)